+++ container_helpers/build_containers.sh
@@ -1,4 +1,4 @@
 #!/bin/sh -e
 mydir="${0%/*}"
 cd $mydir
-docker build -f Dockerfile_docker_run -t docker_run
+docker build -f Dockerfile_docker_run -t localhost/docker_run:latest