+++ res/runit/04-ohx.sh
@@ -6,3 +6,7 @@ mount -o rw,remount / || emergency_shell
 [ "$(grep "polkitd" < /etc/passwd|wc -l)" = "0" ] && useradd polkitd
 [ "$(grep "chrony" < /etc/passwd|wc -l)" = "0" ] && useradd chrony
 mount -o ro,remount / || emergency_shell
+
+msg "Add IP 192.168.4.1/24 to wifi link"
+wifi=$(cat /proc/net/wireless | tail -n1 | cut -d ":" -f1)
+ip addr add 192.168.4.1/24 dev $wifi broadcast 255.255.255.0